tileatlas: split producer webp arm into its own tu + strip redundant test casts
integration fixups for the ereader-finish wave. #290's webp arm pushed
ra8_tileatlas_produce.c to 1143 lines, over the 1000-line file-size cap,
and three of the new host tests carried redundant TEST_ASSERT_EQ casts.

- move the whole-frame webp path (webp_work_bytes, pull_all, feed,
  transcode) into ra8_tileatlas_produce_webp.c and promote the shared
  producer state, prefix-pull, on_geom and on_rows seams into
  ra8_tileatlas_internal.h behind RA8_PRIV, so the dispatcher and the webp
  arm drive one producer. produce.c drops to 829 lines; the emitted atlas
  bytes are unchanged, so the byte-identical webp==png golden is preserved.
- strip_assert_casts on the comic, manga-stream and webp producer tests.

+++ tests/test_ra8_manga_stream.c
@@ -79,7 +79,7 @@ static void t_mg_crosscheck_raw(ra8_vmem_stream_t* st, uint32_t page, uint32_t b
                        (uint64_t)k_ra8_tileatlas_hdr_bytes +
                        ((uint64_t)band * (uint64_t)k_mg_payload);
   const size_t   got = ra8_vmem_stream_read(st, abs, raw, (size_t)k_mg_payload);
-  TEST_ASSERT_EQ((size_t)k_mg_payload, got);
+  TEST_ASSERT_EQ(k_mg_payload, got);
   for (uint32_t i = 0U; i < (uint32_t)k_mg_payload; i += (uint32_t)k_mg_pix_stride) {
     TEST_ASSERT_EQ(t_mg_payload_byte((uint64_t)page, band, i), raw[i]);
   }
